Raise ConnectionError for more clearly what happened. (#13)
* Raise ConnectionError for more clearly what happened.

Currently, 'Tapyrus::RPC.client' throw JSONParseError when response is 401 Unauthorized.
It is a bit confused what happened. So that I implemented ConnectionError. It is raised when response is not 200 OK for cause to be more clearly that client could not connect RPC server.

* Update lib/tapyrus/rpc/tapyrus_core_client.rb

Change name to more correct.

Co-authored-by: shigeyuki azuchi <azuchi@chaintope.com>

* Change name from ConnectionError to Error

Because this Error is raised when RPC Server return error status. so that it is strange to named `ConnectionError`.

* Raise Tapyrus::RPC::Error when server return other than 200 OK

* Change message format to `Hash` for consistency

Before, the format of `message` is either `String` or `Hash`. It is inconvenient for clients.

* Remove `response` field

IMO, `response` field no longer use. The `message` field include response info.

Co-authored-by: shigeyuki azuchi <azuchi@chaintope.com>